Add quotes around printed domain values
The domain intervals can be ambiguous in cases like
SELECT * FROM table WHERE col >= 'a' AND col <= 'b'
vs.
SELECT * FROM table WHERE col = 'a, b'
which both have the domain interval [a, b].
This is fixed by adding quotes to the output (["a", "b"] vs. ["a, b"])
and escaping quotes which may appear in the string itself.
Also made PlanPrinter.formatDomain() use the formatting code in
Range.toString(), which is exactly the same.